Moved similarity index specs in separate folder and fixed type in a file name
